packages/pyxecm/pyproject.toml
[project]
name = "pyxecm"
version = "0.0.0"
description = 'A Python library to interact with Opentext Content Management Rest API'
keywords = [
  'opentext',
  'extendedecm',
  'contentserver',
  'otds',
  'appworks',
  'archivecenter',
]
readme = "README.md"
authors = [
  { name = "Kai Gatzweiler", email = "kgatzweiler@opentext.com" },
  { name = "Dr. Marc Diefenbruch", email = "mdiefenb@opentext.com" },
]
requires-python = '>=3.10'
dependencies = [
  "lxml>=6.0.0",
  "opentelemetry-api>=1.34.1",
  "opentelemetry-exporter-otlp>=1.34.1",
  "opentelemetry-instrumentation-requests>=0.55b1",
  "opentelemetry-instrumentation-threading>=0.55b1",
  "opentelemetry-sdk>=1.34.1",
  "pandas>=2.3.1",
  "requests>=2.32.4",
  "requests-toolbelt>=1.0.0",
  "suds>=1.2.0",
  "websockets>=15.0.1",
  "xmltodict>=0.14.2",
]
classifiers = [
  'Development Status :: 4 - Beta',
  'Programming Language :: Python :: 3',
  'License :: OSI Approved :: Apache Software License',
  'Operating System :: OS Independent',
  'Intended Audience :: Developers',
  'Topic :: Internet :: WWW/HTTP :: Dynamic Content :: Content Management System',
]

[project.optional-dependencies]
magic = ['python-magic']
sap = ['pyrfc==3.3.1']
pyvis = ["pyvis>=0.3.2"]
customizer = [
  "kubernetes>=33.1.0",
  "playwright>=1.53.0",
  "pydantic>=2.11.7",
  "python-hcl2>=7.2.1",
]
api = [
  "asyncio>=3.4.3",
  "fastapi>=0.116.0",
  "jinja2>=3.1.6",
  "opentelemetry-api>=1.34.1",
  "opentelemetry-instrumentation-fastapi>=0.55b1",
  "opentelemetry-sdk>=1.34.1",
  "prometheus-fastapi-instrumentator>=7.1.0",
  "pydantic-settings>=2.10.1",
  "python-multipart>=0.0.20",
  "uvicorn>=0.35.0",
]

[project.urls]
Homepage = 'https://github.com/opentext/pyxecm' 

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
  "src/pyxecm",
  "src/pyxecm_api",
  "src/pyxecm_customizer",
  "src/pyxecm_maintenance_page",
]

[tool.hatch.build.targets.sdist]
exclude = [
  "tests/**"
]

[project.scripts]
pyxecm-api = "pyxecm_api:run_api"
pyxecm-customizer = "pyxecm_customizer:main"
pyxecm-maintenance-page = "pyxecm_maintenance_page:run_api"
